# openapi.yaml
openapi: 3.0.3
info:
  title: Push Balancer API
  description: |
    ML-gestützte API für BILD Push-Notification-Optimierung.

    Der Push Balancer analysiert historische Push-Daten, prognostiziert Opening Rates (OR)
    via Gradient Boosted Trees und Stacking Ensemble (LightGBM + XGBoost + CatBoost → Ridge),
    und liefert tagesaktuelle Empfehlungen für Redaktion und Sport.

    **Safety**: Alle Vorhersagen sind ausschließlich beratend (ADVISORY_ONLY). Das System
    darf niemals autonom Push-Benachrichtigungen versenden. Jede Prediction enthält
    `advisoryOnly: true` und `actionAllowed: false`.
  version: 2.4.0
  contact:
    name: BILD Digital Produkt
    email: push-balancer@axelspringer.com
servers:
  - url: http://localhost:8050
    description: Lokaler Entwicklungsserver
  - url: https://push-balancer.onrender.com
    description: Render Cloud Deployment
...
paths:
...
  /api/ml/status:
    get:
      operationId: getMlStatus
      tags: [ml]
      summary: ML-Modell-Status abrufen
      description: |
        Liefert den aktuellen Status des LightGBM-Modells: ob es trainiert ist,
        Trainings-Metriken (MAE, RMSE, R²), SHAP Feature Importances, Trainings-Zähler
        und Zeitstempel des letzten und nächsten Retrainings.
      responses:
        "200":
          description: ML-Modell-Status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MlStatusResponse"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/ml/safety-status:
    get:
      operationId: getMlSafetyStatus
      tags: [ml]
      summary: ML Safety-Status abrufen
      description: |
        Prüft und gibt den Safety-Status des Systems zurück. Das System läuft
        ausschließlich im `ADVISORY_ONLY`-Modus; `actionAllowed` ist immer `false`.
        Dieser Endpoint ist für Monitoring und Compliance-Checks gedacht.
      responses:
        "200":
          description: Safety-Status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SafetyStatusResponse"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/ml/predict:
    get:
      operationId: getMlPredict
      tags: [ml]
      summary: Einzelne OR-Prognose (LightGBM)
      description: |
        Berechnet eine einzelne Opening-Rate-Prognose (XOR) für einen Push-Artikel
        via LightGBM-Ensemble. Gibt `advisoryOnly: true` zurück — das Ergebnis ist
        rein beratend und darf nicht für automatisches Pushen genutzt werden.
      parameters:
        - name: title
          in: query
          required: false
          description: Push-Titelzeile (max. 120 Zeichen)
          schema:
            type: string
            example: "Merz zum Kanzler gewählt: So reagiert Deutschland"
        - name: cat
          in: query
          required: false
          description: Artikel-Kategorie (z.B. news, sport, politik, unterhaltung)
          schema:
            type: string
            default: news
            example: politik
        - name: hour
          in: query
          required: false
          description: Stunde des geplanten Versands (0–23)
          schema:
            type: integer
            minimum: 0
            maximum: 23
            example: 18
        - name: weekday
          in: query
          required: false
          description: Wochentag (0=Montag … 6=Sonntag)
          schema:
            type: integer
            minimum: 0
            maximum: 6
            example: 2
        - name: eilmeldung
          in: query
          required: false
          description: "1 wenn Eilmeldung/Breaking News, sonst 0"
          schema:
            type: integer
            enum: [0, 1]
            default: 0
      responses:
        "200":
          description: OR-Prognose mit Safety-Envelope
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PredictionResult"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/ml/experiments:
    get:
      operationId: getMlExperiments
      tags: [ml]
      summary: Trainings-Experimente auflisten
      description: |
        Gibt eine Liste aller gespeicherten Trainings-Experimente zurück (Experiment-ID,
        Zeitstempel, Metriken, Hyperparameter), geordnet nach Zeitstempel absteigend.
      responses:
        "200":
          description: Liste der Experimente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MlExperiment"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/ml/experiments/compare:
    get:
      operationId: compareMlExperiments
      tags: [ml]
      summary: Trainings-Experimente vergleichen
      description: |
        Vergleicht zwei oder mehr Trainings-Experimente anhand ihrer Metriken
        (MAE, RMSE, R², Feature Count). Nützlich für Champion/Challenger-Bewertung.
      parameters:
        - name: ids
          in: query
          required: true
          description: Kommagetrennte Liste von Experiment-IDs
          schema:
            type: string
            example: "exp_20260401_143022,exp_20260402_091544"
      responses:
        "200":
          description: Vergleichsergebnis
          content:
            application/json:
              schema:
                type: object
                properties:
                  experiments:
                    type: array
                    items:
                      $ref: "#/components/schemas/MlExperiment"
                  best:
                    type: string
                    description: ID des besten Experiments (niedrigstes MAE)
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/ml/ab-status:
    get:
      operationId: getMlAbStatus
      tags: [ml]
      summary: A/B-Test-Status abrufen
      description: |
        Gibt den aktuellen Status des Champion/Challenger A/B-Tests zurück.
        Enthält ob ein Test aktiv ist, IDs von Champion und Challenger,
        Anzahl der gesammelten Samples und Zeitrahmen.
      responses:
        "200":
          description: A/B-Test-Status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AbStatusResponse"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/ml/monitoring:
    get:
      operationId: getMlMonitoring
      tags: [ml]
      summary: ML-Monitoring-Dashboard abrufen
      description: |
        Liefert umfassendes Monitoring-Dashboard: MAE-Trends (24h, 7d), Calibration
        Bias, A/B-Test-Zusammenfassung, Online-Learning-Status, Feature-Drift-Detektion,
        Drift-State und jüngste Monitoring-Events aus der Datenbank.
      responses:
        "200":
          description: Monitoring-Daten
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MlMonitoringResponse"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/ml/retrain:
    post:
      operationId: triggerMlRetrain
      tags: [ml]
      summary: ML-Modell neu trainieren (manuell)
      description: |
        Startet asynchron ein Neu-Training des LightGBM-Modells im Hintergrund.
        Wenn bereits ein Training läuft, wird `status: "already_training"` zurückgegeben.
        Das Training nutzt alle reifen Pushes aus der SQLite-Datenbank.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Training-Status
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [training_started, already_training]
                    example: training_started
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/predict-batch:
    post:
      operationId: predictBatch
      tags: [ml]
      summary: Batch-OR-Prognose für mehrere Artikel
      description: |
        Berechnet Opening-Rate-Prognosen für bis zu 300 Artikel in einem Request.
        Nutzt einen 5-Minuten-Cache mit Micro-Variation für Live-Ticker-Updates.
        Schneller Pfad: GBRT → LightGBM → Fallback (kein Heuristik-Overhead).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PredictBatchRequest"
      responses:
        "200":
          description: Batch-Prognosen
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PredictBatchResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /api/ml/monitoring/tick:
    post:
      operationId: triggerMonitoringTick
      tags: [ml]
      summary: Monitoring-Tick manuell auslösen
      description: |
        Löst manuell einen Monitoring-Tick aus (MAE-Berechnung, Calibration-Update,
        Drift-Detektion). Normalerweise läuft dies automatisch im Hintergrund.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Tick erfolgreich
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  lastTick:
                    type: number
                    description: Unix-Timestamp des letzten Ticks
                    example: 1743590400
        "500":
          $ref: "#/components/responses/InternalServerError"

# ── Components ────────────────────────────────────────────────────────────
components:
  schemas:
...
    Problem:
      type: object
      description: RFC 9457 Problem Details für HTTP APIs
      required:
        - type
        - title
        - status
      properties:
        type:
          type: string
          format: uri
          description: URI-Referenz, die den Fehlertyp identifiziert
          example: "https://push-balancer.example.com/problems/validation-error"
        title:
          type: string
          description: Kurze, menschenlesbare Fehlerbeschreibung
          example: "Validation Error"
        status:
          type: integer
          description: HTTP-Statuscode
          example: 400
        detail:
          type: string
          description: Ausführlichere Fehlerbeschreibung
          example: "Das Feld 'pushId' darf nicht leer sein"
        instance:
          type: string
          format: uri
          description: URI-Referenz des konkreten Fehler-Vorkommens
          example: "/api/prediction-feedback"

    # ── Health ──────────────────────────────────────────────────────────────
...
    MlStatusResponse:
      type: object
      properties:
        trained:
          type: boolean
          description: Ob das Modell trainiert ist
          example: true
        training:
          type: boolean
          description: Ob gerade ein Training läuft
          example: false
        metrics:
          $ref: "#/components/schemas/MlMetrics"
        shapImportance:
          type: array
          description: Top-10 SHAP Feature Importances
          items:
            $ref: "#/components/schemas/FeatureImportanceEntry"
        trainCount:
          type: integer
          description: Anzahl abgeschlossener Trainings-Zyklen
          example: 12
        lastTrainTs:
          type: number
          description: Unix-Timestamp des letzten Trainings
          example: 1743500000
        nextRetrainTs:
          type: number
          description: Unix-Timestamp des geplanten nächsten Trainings
          example: 1743586400
        featureCount:
          type: integer
          description: Anzahl verwendeter Features
          example: 42
    MlMetrics:
      type: object
      properties:
        mae:
          type: number
          format: float
          description: Mean Absolute Error in Prozentpunkten
          example: 1.42
        rmse:
          type: number
          format: float
          description: Root Mean Squared Error in Prozentpunkten
          example: 2.01
        r2:
          type: number
          format: float
          description: R²-Score (Bestimmtheitsmaß)
          example: 0.71
        trainN:
          type: integer
          description: Anzahl Trainings-Samples
          example: 2800
        valN:
          type: integer
          description: Anzahl Validierungs-Samples
          example: 600
        testN:
          type: integer
          description: Anzahl Test-Samples
          example: 400
    FeatureImportanceEntry:
      type: object
      properties:
        name:
          type: string
          description: Feature-Name
          example: hour_sin
        importance:
          type: number
          format: float
          description: Relative Wichtigkeit (0–1)
          example: 0.187
    SafetyStatusResponse:
      type: object
      required: [safetyMode, advisoryOnly, actionAllowed]
      properties:
        safetyMode:
          type: string
          description: Immer "ADVISORY_ONLY"
          example: ADVISORY_ONLY
        advisoryOnly:
          type: boolean
          description: Immer true
          example: true
        actionAllowed:
          type: boolean
          description: Immer false — das System darf niemals autonom pushen
          example: false
        status:
          type: string
          example: "OK — System ist im ADVISORY_ONLY Modus"
    PredictionResult:
      type: object
      description: OR-Prognose mit Safety-Envelope
      properties:
        predictedOr:
          type: number
          format: float
          description: Prognostizierte Opening Rate in Prozent
          example: 7.24
        or:
          type: number
          format: float
          description: Alias für predictedOr (Kompatibilität)
          example: 7.24
        q10:
          type: number
          format: float
          description: 10%-Quantil der Prognose
          example: 4.8
        q90:
          type: number
          format: float
          description: 90%-Quantil der Prognose
          example: 10.2
        confidence:
          type: number
          format: float
          description: Modell-Konfidenz (0–1)
          example: 0.72
        basis:
          type: string
          description: Modell-Basis (z.B. lgbm, ensemble, fallback)
          example: ensemble
        advisoryOnly:
          type: boolean
          description: Safety-Flag — immer true
          example: true
        actionAllowed:
          type: boolean
          description: Safety-Flag — immer false
          example: false
        safetyMode:
          type: string
          example: ADVISORY_ONLY
    MlExperiment:
      type: object
      properties:
        experimentId:
          type: string
          example: exp_20260401_143022
        timestamp:
          type: number
          description: Unix-Timestamp des Trainings
          example: 1743500000
        metrics:
          $ref: "#/components/schemas/MlMetrics"
        promoted:
          type: boolean
          description: Ob dieses Experiment als Champion promoted wurde
          example: false
        modelType:
          type: string
          example: stacking_ensemble
    AbStatusResponse:
      type: object
      properties:
        active:
          type: boolean
          description: Ob ein A/B-Test aktiv ist
          example: true
        championId:
          type: string
          description: Experiment-ID des Champion-Modells
          example: exp_20260401_143022
        challengerId:
          type: string
          description: Experiment-ID des Challenger-Modells
          example: exp_20260402_091544
        nSamples:
          type: integer
          description: Anzahl bisher gesammelter A/B-Samples
          example: 47
        startedAt:
          type: number
          description: Unix-Timestamp des Test-Starts
          example: 1743500000
        timeoutAt:
          type: number
          description: Unix-Timestamp des geplanten Test-Endes
          example: 1743586400
    MlMonitoringResponse:
      type: object
      properties:
        mae24h:
          type: number
          format: float
          description: MAE der letzten 24 Stunden
          example: 1.38
        mae7d:
          type: number
          format: float
          description: MAE der letzten 7 Tage
          example: 1.54
        maeTrend:
          type: array
          description: Zeitreihe der MAE-Werte
          items:
            type: number
        calibrationBias:
          type: number
          format: float
          description: Globaler Calibration Bias (predicted − actual, Durchschnitt)
          example: 0.12
        calibrationTrend:
          type: array
          items:
            type: number
        abSummary:
          type: object
          properties:
            active:
              type: boolean
            nSamples:
              type: integer
            challengerId:
              type: string
        onlineStatus:
          type: object
          properties:
            paused:
              type: boolean
            updatesCount:
              type: integer
            onlineMae:
              type: number
            batchMae:
              type: number
        driftState:
          type: object
          properties:
            detected:
              type: boolean
              description: Ob Data-Drift erkannt wurde
            currentMae:
              type: number
            historicalMae:
              type: number
            autoRetrainCount:
              type: integer
        recentEvents:
          type: array
          items:
            $ref: "#/components/schemas/MonitoringEvent"
        lastTick:
          type: number
          description: Unix-Timestamp des letzten Monitoring-Ticks
    MonitoringEvent:
      type: object
      properties:
        id:
          type: integer
        timestamp:
          type: number
        eventType:
          type: string
          example: drift_detected
        severity:
          type: string
          enum: [info, warning, error]
          example: warning
        message:
          type: string
          example: "MAE-Anstieg von 1.4 auf 2.1pp erkannt"
        metrics:
          type: object
          additionalProperties: true

    # ── Batch Prediction ────────────────────────────────────────────────────
    PredictBatchArticle:
      type: object
      required: [id]
      properties:
        id:
          type: string
          description: Eindeutige Artikel-ID (für Cache-Lookup)
          example: "msg_12345"
        title:
          type: string
          description: Push-Titelzeile
          example: "Sensation im Finale: Bayern siegt!"
        cat:
          type: string
          description: Kategorie
          example: Sport
        hour:
          type: integer
          description: Versand-Stunde (0–23)
          example: 20
        isEilmeldung:
          type: boolean
          description: Eilmeldung?
          example: false
        isBildPlus:
          type: boolean
          description: BILD-Plus-Artikel?
          example: false
    PredictBatchRequest:
      type: object
      required: [articles]
      properties:
        articles:
          type: array
          description: Liste zu prognostizierender Artikel (max. 300)
          maxItems: 300
          items:
            $ref: "#/components/schemas/PredictBatchArticle"
    PredictBatchPrediction:
      type: object
      properties:
        or:
          type: number
          format: float
          description: Prognostizierte OR (mit Micro-Variation für Live-Ticker)
          example: 6.87
        predictedOr:
          type: number
          format: float
          example: 6.87
        basis:
          type: string
          example: ensemble
        confidence:
          type: number
          format: float
          example: 0.68
        q10:
          type: number
          format: float
          example: 4.2
        q90:
          type: number
          format: float
          example: 9.8
        modelType:
          type: string
          example: ensemble
    PredictBatchResponse:
      type: object
      properties:
        predictions:
          type: object
          description: Map von Artikel-ID zu Prognose
          additionalProperties:
            $ref: "#/components/schemas/PredictBatchPrediction"
        modelType:
          type: string
          description: Verwendetes Modell (aus erstem Ergebnis)
          example: ensemble
        count:
          type: integer
          description: Anzahl berechneter Prognosen
          example: 42
        elapsedMs:
          type: number
          description: Berechnungszeit in Millisekunden
          example: 18.4
        cached:
          type: integer
          description: Anzahl aus dem Cache gelieferter Prognosen
          example: 38

    # ── Prediction Feedback ─────────────────────────────────────────────────
...
  responses:
    BadRequest:
      description: Ungültige Anfrage (fehlende oder falsche Parameter)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: "https://push-balancer.example.com/problems/bad-request"
            title: "Bad Request"
            status: 400
            detail: "Das Feld 'pushId' darf nicht leer sein"
...
    InternalServerError:
      description: Interner Serverfehler
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: "https://push-balancer.example.com/problems/internal-error"
            title: "Internal Server Error"
            status: 500
            detail: "ML status error: model not initialized"
...